Add iteration by Part (data about each line).
Lint was not working correctly, fixed.  This
showed that the RegExp we had for newlines actually
DID have a ReDoS in it, even after the language
that was in the docs.

Added per-call firstCol as well, so you don't have
to create a new instance for every use just for that.

Note that this has a workaround for microsoft/TypeScript#61448.